Removed colorset 3 (pager colors) in favor of just using the default color set.
Modified the inactive colorset to be a gradient with the background color and
highlight back so that it contrasts better with the focused window.

Updated icon used for chrome to match the current icomoon name.

Removed the "Theme" menu in favor of the generated color menus.
Added save option to the custom color scheme.  There is currently no delete
option, edit the `~/.fvwm/.theme` file to delete saved entries.

Fixed spacing in the test for GOG menu population (can't have a space between
the ! and x in Test expressions).

Modified fcitx launch to only happen on initial load.

Power button uses custom quit form instead of the fvwm quit verify.

Moved ReStyle function destruction to before sourcing preference files so that
preference files can modify the ReStyle function.  There are others that will
need to move, too.